requirements.txt
Django==4.2.7
psycopg2-binary
django-environ
pytest-django
pytest-cov
black
ruff
djangorestframework
djangorestframework-simplejwt
weasyprint>=60.0
pandas>=2.0
xlsxwriter>=3.2
drf-spectacular>=0.27
drf-spectacular-sidecar
pyyaml

mypy
